a lot of small stuff; starting to remove ShaderPlus

+++ src/me/anno/ui/editor/sceneView/Gizmos.kt
@@ -75,7 +75,7 @@ object Gizmos {
         }
         local.scale(scale)
         shader.m4x3delta("localTransform", local, camPosition, worldScale)
-        shader.v1f("worldScale", worldScale.toFloat())
+        shader.v1f("worldScale", worldScale)
         material.bind(shader)
         shader.v4f("diffuseBase", color or (255 shl 24))
         shaderColor(shader, "tint", color or (255 shl 24))
